pyproject.toml
[project]
name = "django-docmgr"
version = "0.5.0"
description = "A pluggable Django app to reference documents (files) in models."
keywords = ["django", "documents", "files", "contenttypes", "drf", "storage"]
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Michael Heumann", email="normic@mailbox.org"},
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "License-Expression: BSD-3-Clause",
    "License-File: LICENSE.txt",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
#    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

dependencies = [
    "django>=4.2",
    "django-braces>=1.8.1",
    "django-downloadview>=2.4.0",
]

[project.optional-dependencies]
# Install API support with: pip install django-docmgr[drf]
drf = [
    "djangorestframework>=3.14",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]

[build-system]
requires = ["setuptools>=62", "wheel"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/normic/django-docmgr"
